update : useTimeAgo->timeAgo 일반함수로 변경. 버그 수정됨
