+++ src/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from './useVerifyUser';
 export type { tempUserState } from './useVerifyUser';
 export * from './useLike';
-export * from './useTimeAgo';
